+++ libraries/Student/Student.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: Student.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 // PURPOSE: Arduino library for Student or T-distribution math.
 //    DATE: 2024-07-22
 //     URL: https://github.com/RobTillaart/Student